server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
aiohttp==3.10.10
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
ujson==5.10.0

# Google Cloud dependencies
google-cloud-storage>=2.7.0
google-auth>=2.17.0
python-dotenv>=1.0.0


